checking to improve performance

+++ kcore/block-level-dual-compact/src/host_funcs.cu
@@ -38,7 +38,7 @@ void find_kcore(string data_file,bool write_to_disk){
 
     unsigned int level = 0;
     unsigned int *global_count;
-    volatile unsigned int* blockCounter;
+    volatile unsigned long long int* blockCounter;
     cudaMallocManaged(&global_count,sizeof(unsigned int));
     cudaMallocManaged(&blockCounter,sizeof(unsigned int));
